requirements.txt
# Core Flask dependencies (Python 3.10+ compatible)
Flask>=2.2.5
Werkzeug>=3.1.3

# Database connectivity
pyodbc>=5.2.0
keyring>=25.2.0

# AI/ML services
openai>=2.3.0

# Environment management
python-dotenv>=1.1.1

# Task scheduling
APScheduler>=3.10.4
pytz>=2024.1

# Development dependencies (optional)
pytest>=8.4.2
pytest-flask>=1.3.0
black>=21.12b0
flake8>=4.0.1
mypy>=0.910

# Production server (optional)
gunicorn>=23.0.0

# Windows production server (optional)
waitress>=2.1.2

# Monitoring & structured logging (optional)
prometheus-client>=0.20.0
structlog>=24.4.0

# Windows keyring backend helpers (Windows only; usually pulled transitively)
pywin32-ctypes>=0.2.2